requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
faiss-cpu==1.8.0
sentence-transformers==2.2.2
groq==0.4.1
python-dotenv==1.0.0